SESSION IS SET FOR THE API

+++ src/main/java/personalfinancetrackerinweb/api/BudgetRestApi.java
@@ -17,63 +17,85 @@ import personalfinancetrackerinweb.model.Budget;
 import personalfinancetrackerinweb.model.User;
 import personalfinancetrackerinweb.repository.BudgetRepositoryImpl;
 
-
 @Path("/budgetSource")
 
 @Produces(MediaType.APPLICATION_JSON)
 @Consumes(MediaType.APPLICATION_JSON)
-public class BudgetRestApi implements Serializable {
+public class BudgetRestApi extends BaseRestApi implements Serializable {
 
     @Inject
     private BudgetRepositoryImpl budgetRepositoryImpl;
-    
-    private User user;
 
-    public User getUser() {
-        return user;
-    }
-    
     @GET
     public Response getAllBudgets() {
-        List<Budget> budgets = budgetRepositoryImpl.findByUser(1);
-        RestResponse responseModel = new RestResponse("true",200, "Success", budgets);
-        return Response.status(200).entity(responseModel).build();
+        User loggedInUser = getLoggedInUser();
+        if (loggedInUser != null) {
+            List<Budget> budgets = budgetRepositoryImpl.findByUser(loggedInUser.getId());
+            RestResponse responseModel = new RestResponse("true", 200, "Success", budgets);
+            return Response.status(200).entity(responseModel).build();
+        } else {
+            RestResponse responseModel = new RestResponse("false", 401, "Unauthorized", null);
+            return Response.status(401).entity(responseModel).build();
+        }
     }
 
     @GET
     @Path("/{id}")
     public Response getBudgetById(@PathParam("id") long id) {
-        Budget budget = budgetRepositoryImpl.getById(id);
-        if (budget != null) {
-            RestResponse responseModel = new RestResponse("true",200, "Success", budget);
-            return Response.status(200).entity(responseModel).build();
+        User loggedInUser = getLoggedInUser();
+        if (loggedInUser != null) {
+            Budget budget = budgetRepositoryImpl.getById(id);
+            if (budget != null) {
+                RestResponse responseModel = new RestResponse("true", 200, "Success", budget);
+                return Response.status(200).entity(responseModel).build();
+            } else {
+                RestResponse responseModel = new RestResponse("false", 404, "Budget not found", null);
+                return Response.status(404).entity(responseModel).build();
+            }
         } else {
-            RestResponse responseModel = new RestResponse("false",404, "Budget not found", null);
-            return Response.status(404).entity(responseModel).build();
+            RestResponse responseModel = new RestResponse("false", 401, "Unauthorized", null);
+            return Response.status(401).entity(responseModel).build();
         }
     }
 
     @POST
     public Response createBudget(Budget budget) {
-        budgetRepositoryImpl.create(budget);
-        RestResponse responseModel = new RestResponse("true",201, "Data Saved successfully!", budget);
-        return Response.status(201).entity(responseModel).build();
+        User loggedInUser = getLoggedInUser();
+        if (loggedInUser != null) {
+            budgetRepositoryImpl.create(budget);
+            RestResponse responseModel = new RestResponse("true", 201, "Data Saved successfully!", budget);
+            return Response.status(201).entity(responseModel).build();
+        } else {
+            RestResponse responseModel = new RestResponse("false", 401, "Unauthorized", null);
+            return Response.status(401).entity(responseModel).build();
+        }
     }
 
     @PUT
     @Path("/{id}")
     public Response updateBudget(@PathParam("id") long id, Budget budget) {
-        budgetRepositoryImpl.update(budget);
-        RestResponse responseModel = new RestResponse("true",200, "Data Updated Successfully!", budget);
-        return Response.status(200).entity(responseModel).build();
+        User loggedInUser = getLoggedInUser();
+        if (loggedInUser != null) {
+            budgetRepositoryImpl.update(budget);
+            RestResponse responseModel = new RestResponse("true", 200, "Data Updated Successfully!", budget);
+            return Response.status(200).entity(responseModel).build();
+        } else {
+            RestResponse responseModel = new RestResponse("false", 401, "Unauthorized", null);
+            return Response.status(401).entity(responseModel).build();
+        }
     }
 
     @DELETE
     @Path("/{id}")
-    
     public Response deleteBudget(@PathParam("id") long id) {
+        User loggedInUser = getLoggedInUser();
+        if (loggedInUser != null) {
         budgetRepositoryImpl.delete(id);
-        RestResponse responseModel = new RestResponse("true",202, "Data Deleted Successfully!", null);
+        RestResponse responseModel = new RestResponse("true", 202, "Data Deleted Successfully!", null);
         return Response.status(202).entity(responseModel).build();
-    }
+        }else{
+           RestResponse responseModel = new RestResponse("false", 401, "Unauthorized", null);
+            return Response.status(401).entity(responseModel).build(); 
+        }
+   }
 }
